pyproject.toml
[project]
name = "youtube-university"
version = "0.1.0"
description = "Extract actionable knowledge from YouTube channels using AI"
requires-python = ">=3.9"
dependencies = [
    "youtube-transcript-api>=1.2.4",
    "requests>=2.28.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "flask>=3.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.0",
]

[project.scripts]
ytuni = "youtube_university.cli:cli"

[build-system]
requires = ["setuptools>=64.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["youtube_university*"]

[tool.ruff]
line-length = 100
target-version = "py39"
